// packages/next-js/src/lib/json-schema-types.ts
export interface JSONSchema {
	$schema?: string;
	$id?: string;
	title?: string;
	description?: string;
	type?: string | string[];
	properties?: Record<string, JSONSchema>;
	required?: string[];
	additionalProperties?: boolean | JSONSchema;
	items?: JSONSchema | { anyOf: JSONSchema[] } | { oneOf: JSONSchema[] };
	enum?: (string | number | boolean | null)[];
	const?: unknown;
	default?: unknown;
	examples?: unknown[];
	format?: string;
	pattern?: string;
	minLength?: number;
	maxLength?: number;
	minimum?: number;
	maximum?: number;
	exclusiveMinimum?: number | boolean;
	exclusiveMaximum?: number | boolean;
	multipleOf?: number;
	minItems?: number;
	maxItems?: number;
	uniqueItems?: boolean;
	minProperties?: number;
	maxProperties?: number;
	allOf?: JSONSchema[];
	anyOf?: JSONSchema[];
	oneOf?: JSONSchema[];
	not?: JSONSchema;
	if?: JSONSchema;
	then?: JSONSchema;
	else?: JSONSchema;
	$ref?: string;
	definitions?: Record<string, JSONSchema>;
	$defs?: Record<string, JSONSchema>;
	deprecated?: boolean;
	readOnly?: boolean;
	writeOnly?: boolean;
}